nfc: fix buffer overread in MIFARE-Classic TLV parser, harden bounds checks & add unit tests
• get_mifare_classic_ndef_start_index()
  ▸ return type changed to int8_t and explicit cast in loop
  ▸ loop capped with (int8_t)MIFARE_CLASSIC_BLOCK_SIZE to silence -Wsign-compare
  ▸ guarantees return range −2…15 (documented)

• decode_mifare_classic_tlv()
  ▸ validates start index before dereference
  ▸ verifies that short/long TLV header fits into buffer
  ▸ casts to uint16_t when assigning to message_start_index

• All PN532 / PN7150 / PN7160 call sites updated to new signature

**Tests**
  ▸ pytest/hypothesis suite:
     – short/long TLV (254 B, 255 B)
     – terminator case, off-by-one regression (size 300 → 320)
     – fuzz (16-32 B random buffers, 1 000 cases)
     – round-trip short (1-32 B) & long (300 B)
     – negative-path start-index checks

+++ esphome/components/pn532/pn532_mifare_classic.cpp
@@ -10,7 +10,7 @@ static const char *const TAG = "pn532.mifare_classic";
 
 std::unique_ptr<nfc::NfcTag> PN532::read_mifare_classic_tag_(std::vector<uint8_t> &uid) {
   uint8_t current_block = 4;
-  uint8_t message_start_index = 0;
+  uint16_t message_start_index = 0;
   uint32_t message_length = 0;
 
   if (this->auth_mifare_classic_block_(uid, current_block, nfc::MIFARE_CMD_AUTH_A, nfc::NDEF_KEY)) {
